Move interception filter to its own class
Move the IHttpMessageHandlerBuilderFilter for JustEat.HttpClientInterception to its own class.
